src/saaga_mcp_servers/utils/mcp_gemini/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp_gemini"
version = "0.1.0"
description = "An MCP-compatible server implementation"
readme = "README.md"
authors = [
    {name = "Andrew", email = "andrew@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11,<3.13"
dependencies = [
    "mcp>=1.0.0",
    "mcp[cli]>=1.0.0",
    "anyio>=4.5",
    "starlette>=0.36.0",
    "uvicorn>=0.27.0",
    "google-genai>=0.1.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
# Single entry point for server that handles both transports
mcp_gemini-server = "mcp_gemini.server.app:main"
mcp_gemini-client = "mcp_gemini.client.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_gemini*"]
namespaces = false 
